separate the data access logic from the business logic into new layers(folders)
