+++ backend/src/lambda/http/getUserIds.ts
@@ -1,7 +1,7 @@
 import {APIGatewayProxyResult} from 'aws-lambda'
 import * as middy from 'middy'
 import { cors, httpErrorHandler } from 'middy/middlewares'
-import { getUserIds as getUserIds } from '../../helpers/todos'
+import { getUserIds as getUserIds } from '../../businessLogic/todos'
 
 // TODO: Get all TODO items for a current user
 export const handler = middy(
